feat: address todo #18334
