+++ hudi-spark-datasource/hudi-spark4.0.x/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/Spark40ParquetReader.scala
@@ -206,7 +206,7 @@ class Spark40ParquetReader(enableVectorizedReader: Boolean,
       }
     } else {
       // ParquetRecordReader returns InternalRow
-      val readSupport = new HoodieParquetReadSupport(
+      val readSupport = new Spark40HoodieParquetReadSupport(
         convertTz,
         enableVectorizedReader = false,
         enableLogicalTimestampRepair,
